chatgpt/pipeline_generate/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jack"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
langchain = "^0.0.191"
openai = "^0.27.8"
python-consul = "^1.1.0"
grpcio-tools = "^1.57.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
